added a public view for suggestions
