+++ config/environments/production.rb
@@ -88,6 +88,7 @@ Rails.application.configure do
     logger.formatter = config.log_formatter
     config.logger    = ActiveSupport::TaggedLogging.new(logger)
   end
+  config.serve_static_assets = true
 
   # Do not dump schema after migrations.
   config.active_record.dump_schema_after_migration = false
